backend/requirements.txt
# ===============================
# Core web framework & utilities
# ===============================
fastapi~=0.115
uvicorn[standard]~=0.34.0
python-dotenv~=1.1.0
pydantic~=2.11
pydantic-settings~=2.9.0

# ===============================
# Async io, caching, DB, auth
# ===============================
aiohttp~=3.11
aiocache~=0.12.3
aioredis==2.0.1
aioboto3~=14.3
python-jose[cryptography]>=3.3.0

beanie~=1.29.0
motor==3.7.0
pymongo==4.13.0

# ===============================
# HTTP client
# ===============================
httpx~=0.28.0 

# ===============================
# Blockchain SDKs & helpers
# ===============================
# EVM & generic
web3~=7.11.0      # Ethereum, Polygon, BSC та сумісні
eth-typing~=5.2.0
eth-utils~=5.3.0

# Tron
tronpy~=0.5.0

# Solana + SPL
solana~=0.36.0           # Solana Python SDK


# TON
tonutils[pytoniq]~=0.4.0       # Python wrapper for ton-client (crypto & RPC)

# Utilities
base58>=2.1.1
bip-utils~=2.9.3